exposed-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
botocore==1.34.0
pydantic==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
